requirements.txt
numpy==1.19.2
numpy==1.21.6
pandas==1.3.5
Pillow==9.3.0
Pillow==9.5.0
Pillow==12.0.0
ResNet==0.1
scikit_learn==1.0.2
scipy==1.7.3
submitit==1.5.3
tensorboardX==2.2
tensorboardX==2.6.4
timm==0.3.2
torch==1.13.1+cu116
torchlars==0.1.2
torchvision==0.14.1+cu116
tqdm==4.64.0
tqdm==4.65.0
ViT==2.3.3
warmup_scheduler==0.3
